AI_Folder
This is a project I have done during my AI specialization. It uses machine learning models to predict the house prices in Amsterdam, based on the properties of these houses. Additionally a Power BI dashboard was made to visualize the results of the model. Furthermore, more information are added like the notebook, proposal and the final report.